# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A wrapper around ffmpeg",
        "title": "ffmate API",
        "contact": {
            "name": "We love media",
            "email": "sev@welovemedia.io"
        },
        "license": {
            "name": "AGPL-3.0",
            "url": "https://opensource.org/license/agpl-v3"
        }
    },
    "host": "localhost",
    "basePath": "/api/v1",
    "paths": {
...
        "/tasks": {
            "get": {
                "description": "List all existing tasks",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "List all tasks",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "the page of a pagination request (min 0)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "the amount of results of a pagination request (min 1; max: 100)",
                        "name": "perPage",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.Task"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Add a new tasks to the queue",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Add a new task",
                "parameters": [
                    {
                        "description": "new task",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.NewTask"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Task"
                        }
                    }
                }
            }
        },
        "/tasks/batch": {
            "post": {
                "description": "Add a batch of new tasks to the queue",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Add a batch of tasks",
                "parameters": [
                    {
                        "description": "new tasks",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.NewTask"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.Task"
                            }
                        }
                    }
                }
            }
        },
        "/tasks/batch/{uuid}": {
            "get": {
                "description": "Get tasks by batch uuid",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Get tasks for batch",
                "parameters": [
                    {
                        "type": "string",
                        "description": "the batch uuid",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.Task"
                            }
                        }
                    }
                }
            }
        },
        "/tasks/{uuid}": {
            "get": {
                "description": "Get a single task by its uuid",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Get single task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "the tasks uuid",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Task"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a task by its uuid",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Delete a task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "the tasks uuid",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/tasks/{uuid}/cancel": {
            "patch": {
                "description": "Cancel a task by its uuid",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Cancel a task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "the tasks uuid",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Task"
                        }
                    }
                }
            }
        },
        "/tasks/{uuid}/restart": {
            "patch": {
                "description": "Restart a task by its uuid",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Restart a task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "the tasks uuid",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Task"
                        }
                    }
                }
            }
        },
...
        "/webhooks": {
            "get": {
                "description": "List all existing webhooks",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "webhooks"
                ],
                "summary": "List all webhooks",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.Webhook"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Add a new webhook for an event",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "webhooks"
                ],
                "summary": "Add a new webhook",
                "parameters": [
                    {
                        "description": "new webhook",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.NewWebhook"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Webhook"
                        }
                    }
                }
            }
        },
        "/webhooks/{uuid}": {
            "delete": {
                "description": "Delete a webhook by its uuid",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "webhooks"
                ],
                "summary": "Delete a webhook",
                "parameters": [
                    {
                        "type": "string",
                        "description": "the webhooks uuid",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.InterfaceMap": {
            "type": "object",
            "additionalProperties": true
        },
        "dto.NewPrePostProcessing": {
            "type": "object",
            "properties": {
                "scriptPath": {
                    "type": "string"
                },
                "sidecarPath": {
                    "type": "string"
                }
            }
        },
...
        "dto.NewTask": {
            "type": "object",
            "properties": {
                "command": {
                    "type": "string"
                },
                "inputFile": {
                    "type": "string"
                },
                "metadata": {
                    "description": "Additional metadata for the task",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.InterfaceMap"
                        }
                    ]
                },
                "name": {
                    "type": "string"
                },
                "outputFile": {
                    "type": "string"
                },
                "postProcessing": {
                    "$ref": "#/definitions/dto.NewPrePostProcessing"
                },
                "preProcessing": {
                    "$ref": "#/definitions/dto.NewPrePostProcessing"
                },
                "preset": {
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                }
            }
        },
...
        "dto.NewWebhook": {
            "type": "object",
            "properties": {
                "event": {
                    "$ref": "#/definitions/dto.WebhookEvent"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "dto.PrePostProcessing": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "finishedAt": {
                    "type": "integer"
                },
                "scriptPath": {
                    "$ref": "#/definitions/dto.RawResolved"
                },
                "sidecarPath": {
                    "$ref": "#/definitions/dto.RawResolved"
                },
                "startedAt": {
                    "type": "integer"
                }
            }
        },
...
        "dto.RawResolved": {
            "type": "object",
            "properties": {
                "raw": {
                    "type": "string"
                },
                "resolved": {
                    "type": "string"
                }
            }
        },
        "dto.Task": {
            "type": "object",
            "properties": {
                "batch": {
                    "type": "string"
                },
                "command": {
                    "$ref": "#/definitions/dto.RawResolved"
                },
                "createdAt": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                },
                "finishedAt": {
                    "type": "integer"
                },
                "inputFile": {
                    "$ref": "#/definitions/dto.RawResolved"
                },
                "metadata": {
                    "description": "Additional metadata for the task",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.InterfaceMap"
                        }
                    ]
                },
                "name": {
                    "type": "string"
                },
                "outputFile": {
                    "$ref": "#/definitions/dto.RawResolved"
                },
                "postProcessing": {
                    "$ref": "#/definitions/dto.PrePostProcessing"
                },
                "preProcessing": {
                    "$ref": "#/definitions/dto.PrePostProcessing"
                },
                "priority": {
                    "type": "integer"
                },
                "progress": {
                    "type": "number"
                },
                "remaining": {
                    "type": "number"
                },
                "source": {
                    "type": "string"
                },
                "startedAt": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/dto.TaskStatus"
                },
                "updatedAt": {
                    "type": "integer"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "dto.TaskStatus": {
            "type": "string",
            "enum": [
                "QUEUED",
                "RUNNING",
                "PRE_PROCESSING",
                "POST_PROCESSING",
                "DONE_SUCCESSFUL",
                "DONE_ERROR",
                "DONE_CANCELED"
            ],
            "x-enum-varnames": [
                "QUEUED",
                "RUNNING",
                "PRE_PROCESSING",
                "POST_PROCESSING",
                "DONE_SUCCESSFUL",
                "DONE_ERROR",
                "DONE_CANCELED"
            ]
        },
...
        "dto.Webhook": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "event": {
                    "$ref": "#/definitions/dto.WebhookEvent"
                },
                "updatedAt": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "dto.WebhookEvent": {
            "type": "string",
            "enum": [
                "batch.created",
                "batch.finished",
                "task.created",
                "task.updated",
                "task.deleted",
                "preset.created",
                "preset.updated",
                "preset.deleted",
                "webhook.created",
                "webhook.deleted",
                "watchfolder.created",
                "watchfolder.updated",
                "watchfolder.deleted"
            ],
            "x-enum-varnames": [
                "BATCH_CREATED",
                "BATCH_FINISHED",
                "TASK_CREATED",
                "TASK_UPDATED",
                "TASK_DELETED",
                "PRESET_CREATED",
                "PRESET_UPDATED",
                "PRESET_DELETED",
                "WEBHOOK_CREATED",
                "WEBHOOK_DELETED",
                "WATCHFOLDER_CREATED",
                "WATCHFOLDER_UPDATED",
                "WATCHFOLDER_DELETED"
            ]
        }
    }
}